Cambiato tipo di discount, da float a integer

+++ database/migrations/2021_06_19_172113_create_offers_table.php
@@ -18,7 +18,7 @@ class CreateOffersTable extends Migration
             $table->foreignId('event_id');
             $table->date('start');
             $table->date('end');
-            $table->float('discount');
+            $table->integer('discount');
             $table->timestamps();
         });
     }
